+++ diffusion_planner/grpo_run.py
@@ -14,32 +14,7 @@ import sys
 from datetime import datetime
 from pathlib import Path
 
-NCCL_ENV = {
-    "NCCL_NVLS_ENABLE": "0",
-    "NCCL_P2P_DISABLE": "0",
-    "NCCL_IB_DISABLE": "1",
-    "NCCL_SOCKET_IFNAME": "lo",
-    "NCCL_DEBUG": "INFO",
-}
-
-
-def gpu_count() -> int:
-    out = subprocess.run(["nvidia-smi", "-L"], capture_output=True, text=True).stdout
-    return len([ln for ln in out.splitlines() if ln.strip()])
-
-
-def tee_run(cmd: list[str], cwd: Path, env: dict, log_path: Path) -> int:
-    """Run cmd streaming combined stdout/stderr to the console AND log_path (like `2>&1 | tee`)."""
-    with open(log_path, "wb") as log:
-        proc = subprocess.Popen(
-            cmd, cwd=cwd, env=env, stdout=subprocess.PIPE, stderr=subprocess.STDOUT
-        )
-        for line in proc.stdout:
-            sys.stdout.buffer.write(line)
-            sys.stdout.buffer.flush()
-            log.write(line)
-            log.flush()
-        return proc.wait()
+from run_utils import NCCL_ENV, gpu_count, tee_run
 
 
 def parse_args() -> argparse.Namespace:
